feat(projects): Samples and starters (#33)
* feat(projects): Samples and starters

- Add create, add, world, info CLI commands

* feat(projects): Improved drag and drop for files

* fix(editor): Update API references

* fix(gallery): Improve sample relevance, add robustness

- Remove unneeded code

* fix(various): Test and command improvements

- Remove add for now

* fix(samples): Fix some samples provisioning issues

* fix(editor): CreationType update fixes

* fix(editor): Small updates to UI

* feat(project): Split out mobs in gallery

* fix(projects): Remove unused items

- Remove directory picker on platforms that don't support it
- Make command versions a recommendation, not an error

+++ app/src/test/CartoTest.ts
@@ -17,6 +17,7 @@ import ProjectExporter from "../app/ProjectExporter";
 import { IWorldSettings } from "../minecraft/IWorldSettings";
 import { GameType, Generator } from "../minecraft/WorldLevelDat";
 import * as fs from "fs";
+import Utilities from "../core/Utilities";
 
 CartoApp.hostType = HostType.testLocal;
 
@@ -134,7 +135,8 @@ function removeResultFolder(scenarioName: string) {
     const path =
       StorageUtilities.ensureEndsWithDelimiter(resultsFolder.fullPath) +
       StorageUtilities.ensureEndsWithDelimiter(scenarioName);
-    if (fs.existsSync(path))
+    // guard against being called at a "more root" file path
+    if (fs.existsSync(path) && Utilities.countChar(path, NodeStorage.folderDelimiter) > 5)
       // @ts-ignore
       fs.rmSync(path, {
         recursive: true,
